Added Memblocks, fixed bit shift operator precedence and hex integer literal parsing. Added Wait and Chr runtime functions. Fixed putpixel(int, int, int) overload. Added coolnes CBCompiler port to tests.

+++ Runtime/cb_math.cpp
@@ -99,11 +99,11 @@ void CBF_randomize(int seed) {
 	randomize(seed);
 }
 
-float CBF_roundUp(float f) {
+int CBF_roundUp(float f) {
 	return ceilf(f);
 }
 
-float CBF_roundDown(float f) {
+int CBF_roundDown(float f) {
 	return floorf(f);
 }
